Revise README and enhance design with new features and updates
* redesigned UI/UX of the application
* added multi-user system and authentication logic
* implemented shared books collection counter
* removed limitation of single-user workflow
* improved application structure and internal logic
* added project report/documentation

This merge replaces the initial prototype version with a more complete and scalable implementation.
---------

+++ project/db/init.sql
@@ -1,4 +1,3 @@
-
 create table users 
 (
     id serial primary key,
@@ -18,7 +17,7 @@ create table books
     publisher text,
     pages integer,
     isbn text,
-    rating integer,
+    rating NUMERIC(3,1),
     description text,
     created_by integer,
     foreign key (created_by) references users(id) on delete set null
